pyproject.toml
[project]
name = "spending-app-telegram-bot"
version = "0.0.1"
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.11.3",
    "python-dotenv>=1.1.0",
    "python-telegram-bot>=22.0",
]
